function to add event listeners create
